Set up Prog3 enviroment, description provides more details
Set up enviroment for Prog3

main.c holds only the main file
index.c holds only the indexing function
browse.c holds only the browser function
other.c holds all other functions we find necessary

All c and h files besides main.c are empty as of now.

makefile is functional, including a clean target. The only problem is
that when run it gives the error "Error 2 (ignored)" Do not understand
what this means, or if it is a problem, will research this.